Properly handle temporary directory paths
In epub2txt_do_file(), the environment variables TMP and TMPDIR are
used as a base to construct the path for a temporary directory to
extract the EPUB contents into. A long enough value for TMP or TMPDIR
will overrun the fixed-size buffers allocated on the stack of the
function and break the program.

Dynamically allocate strings to avoid buffer overruns, and return early
when mkdir() fails, as there isn't much else to do.