Use execvp instead of system to run helper commands
While easier to use, system() has the disadvantage of needing proper
escaping and validation on the passed parameter, which is passed
as is to a shell, in order to avoid malformed commands and possibly
execution of unwanted commands (command injection).

Add an utility function for running helper commands that uses fork()
+ execvp() + wait() instead of system(). Using execvp() has the
advantage of not needing any escaping on the parameters, and the
interface is also easier to use as ther is no need to construct
commands as a single string using sprintf() or similar functions.